+++ components/TabMenu/index.js
@@ -1,5 +1,6 @@
 import React, { PropTypes } from 'react';
 import ReactDom from 'react-dom';
+import classnames from 'classnames';
 import style from './style.css';
 import TabLink from './TabLink';
 import TabDropdown from './TabDropdown';
@@ -128,32 +129,36 @@ class TabMenu extends React.Component {
             offsetStyle = this.state.offset + 'px';
         }
         return (
-            <div className={this.getStyle('tabMenu')}>
-            <div className={this.getStyle('tabSelectorWrapper')} style={{display: displayBtnStyle}}>
-                <TabDropdown data={this.props.tabs} activeItem={this.state.activeTabIndex} onSelectItem={this.props.onClick} />
-            </div>
-            <table ref='tabset' className={this.getStyle('tabNavbar')} style={{left: offsetStyle}}>
-                <tbody>
-                    <tr>
-                    {this.props.tabs.map((tab, i) => {
-                        let onClose = (e) => {
-                            e.stopPropagation();
-                            e.preventDefault();
-                            let prev = this.props.tabs[i - 1]; // TODO: check perf
-                            let next = this.props.tabs[i + 1]; // TODO: check perf
-                            this.props.onTabClose(tab, prev, next);
-                        };
-                        let handleClick = () => this.props.onClick(tab);
-                        let isActive = tab.id === this.props.activeTab;
-                        return (
-                            <td key={i}>
-                                <TabLink onClose={onClose} {...tab} onClick={handleClick} isActive={isActive} />
-                            </td>
-                        );
-                    })}
-                    </tr>
-                </tbody>
-            </table>
+            <div className={classnames(style.TabMenu, this.getStyle('tabMenu'))}>
+                <div className={this.getStyle('tabSelectorWrapper')} style={{display: displayBtnStyle}}>
+                    <TabDropdown data={this.props.tabs} activeItem={this.state.activeTabIndex} onSelectItem={this.props.onClick} />
+                </div>
+                <div className={style.relativeWrapper}>
+                    <div className={style.absoluteWrapper}>
+                        <table ref='tabset' className={this.getStyle('tabNavbar')} style={{left: offsetStyle}}>
+                            <tbody>
+                                <tr>
+                                    {this.props.tabs.map((tab, i) => {
+                                        let onClose = (e) => {
+                                            e.stopPropagation();
+                                            e.preventDefault();
+                                            let prev = this.props.tabs[i - 1]; // TODO: check perf
+                                            let next = this.props.tabs[i + 1]; // TODO: check perf
+                                            this.props.onTabClose(tab, prev, next);
+                                        };
+                                        let handleClick = () => this.props.onClick(tab);
+                                        let isActive = tab.id === this.props.activeTab;
+                                        return (
+                                            <td key={i}>
+                                                <TabLink onClose={onClose} {...tab} onClick={handleClick} isActive={isActive} />
+                                            </td>
+                                        );
+                                    })}
+                                </tr>
+                            </tbody>
+                        </table>
+                    </div>
+                </div>
             </div>
         );
     }