+++ pages/Master/index.js
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 import ErrorWindow from '../../components/ErrorWindow';
 import Loader from '../../components/Loader';
 import { close } from './actions';
+import style from './style.css';
 
 const ErrorWindowContainer = connect(
     (state) => {
@@ -16,7 +17,7 @@ const Main = ({
     children
 }) => {
     return (
-        <div className='wrapper'>
+        <div className={style.h100pr}>
             { children }
             { loadInfo && !!loadInfo.open && <Loader loadInfo={loadInfo} />}
             <ErrorWindowContainer />